seo-rag/requirements.txt
langchain
pandas
numpy
pyarrow
matplotlib
seaborn
streamlit
faiss-cpu
sentence-transformers
scikit-learn
tqdm
llama-cpp-python
pinecone-client
jupyter
openpyxl
